backend/requirements/dev.txt
-r base.txt

# Debugging
django-debug-toolbar>=4.3,<5.0

# Formatting & linting
black>=24.0,<25.0
isort>=5.13,<6.0
flake8>=7.0,<8.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
factory-boy>=3.3,<4.0

# Type checking
mypy>=1.8,<2.0
django-stubs>=5.0,<6.0
